IGDBService/API/views.py
import requests

from rest_framework.response import Response
from rest_framework.views import APIView

from IGDBService.importdata.models import IGDBGame
from IGDBService.importdata.serializers import IGDBGameSerializer
