requirements.txt
Django==2.1.1
psycopg2
coverage
coveralls
djongo
djangorestframework==3.8.2
